Update data.py, still to be improved before merge
Handling the additional data attributes with conditionals. This is
probably the easiest thing to do for the moment. However, there are
unused (decay_rate has to be removed at the src level) and redundant
(min_depth and bouss_trans_low) parameters that have to be handled
correctly. Others have to be renamed: For instance, bouss_EDC_c_sq
should be replaced by something like bouss_hyprel_param.